[WCNCR00262566] wlan: fix casting (3/.)
[Description]
Fix casting for other system/compiler

MTK-Commit-Id: 40c30d01755e80467365c132c28505aa7d28f5bb

CR-Id: WCNCR00262566
Feature: ${module/feature}
Reviewed-on: https://gerrit.mediatek.inc/c/neptune/wlan_driver/gen4m/+/6106447
Tested-by: yuchi.wang <yuchi.wang@mediatek.com>
CheckCourse: srv_pis <srv_pis@mediatek.com>
CheckCRInfo: srv_pis <srv_pis@mediatek.com>
CheckForbiddenKeyword: srv_pis <srv_pis@mediatek.com>
CheckPatch: srv_pis <srv_pis@mediatek.com>
Build: srv_preflight_a001 <srv_preflight_a001@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity

+++ mgmt/p2p_ie.c
@@ -160,8 +160,8 @@ void p2pGenerate_IEForAssocReq(IN struct ADAPTER *prAdapter,
 
 		prConnReqInfo = &(prP2pRoleFsmInfo->rConnReqInfo);
 
-		pucIEBuf = (uint8_t *) ((unsigned long) prMsduInfo->prPacket
-			+ (unsigned long) prMsduInfo->u2FrameLength);
+		pucIEBuf = (uint8_t *) ((uintptr_t) prMsduInfo->prPacket
+			+ prMsduInfo->u2FrameLength);
 
 		kalMemCopy(pucIEBuf, prConnReqInfo->aucIEBuf,
 			prConnReqInfo->u4BufLength);
@@ -231,8 +231,8 @@ wfdFuncAppendAttriDevInfo(IN struct ADAPTER *prAdapter,
 			break;
 		}
 
-		pucBuffer = (uint8_t *) ((unsigned long) pucBuf
-			+ (unsigned long) (*pu2Offset));
+		pucBuffer = (uint8_t *) ((uintptr_t) pucBuf
+			+ *pu2Offset);
 
 		ASSERT_BREAK(pucBuffer != NULL);
